Update Steam Service and Create Add Steam Appid

+++ src/infrastructure/db/models/user_models.py
@@ -19,7 +19,7 @@ class Users(Base):
     __tablename__ = 'users'
 
     id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
-    steam_id: Mapped[Optional[int]] = mapped_column(Integer, default=None, nullable=True)
+    steam_id: Mapped[Optional[int]] = mapped_column(BigInteger, default=None, nullable=True)
     role: Mapped[str] = mapped_column(String, default="user")
 
     subscribes: Mapped[List["Subscribes"]] = relationship(
